Move data_pipeline to utils.
